Simple formation movement using flowField
Update the flow field to allow multi-source, meaning we can have multiple
destinations in the field that count as being a place where we can stop walking.

When we have reached the formation square following the flow field, we simply
walk to our position in the formation, and turn to face with our squad.