Fix GridPlane setup
First problem was the size of the grid was set to the scale of the child
gameobject, which did not work anymore for the new ground mesh.
Then setting the Grid on the same object as the mesh had the problem of
inheriting the scale and rotatation, causing more out of bounds grid positions.

Created the floor setup to have the navmesh, and gridplane, then have separate
child gameobjects with a mesh, and another for the grid. Scale and rotation and
position offset all contained in the mesh object, not influencing the parent or
sibling.

Fixed the size determination of the grid based on the collider, so that all grid
positions are in the positive space, allowing indexing in the entities array[,]